http: moving the codec filter to its own file
